Move DataProviderComment class to comment directory
